vim: install NeoBundle
I've never fully gotten on board with pathogen or vundle before,
so it's a fresh slate.  And while I like the featureset of vundle
over pathogen, vundle's maintainer stepped down (and the new
maintainers don't seem to have done much yet), while NeoBundle
(which is a fork of vundle) seems to have a clear future.